fix: use long for matrix size, remove nullable returns, and improve dictionary lookup
- Change EditorContentsResponse.Size from int to long to prevent overflow
- Use outputMap.GetSize() and fetchMeta.GetSize() (long) instead of manual LINQ product aggregation
- Add fetchSize > maxQuerySize guard alongside outputMapSize check
- Remove nullable annotations from BuildVirtualValueMaps return tuple
- Replace double dictionary indexer in VirtualValueValidationService with TryGetValue
- Add missing Constant = 0.0 initializer to SumDefinition test fixtures

+++ PxGraf/Models/Responses/EditorContentsResponse.cs
@@ -13,7 +13,7 @@ namespace PxGraf.Models.Responses
         /// <summary>
         /// Current size of the query.
         /// </summary>
-        public required int Size { get; set; }
+        public required long Size { get; set; }
 
         /// <summary>
         /// The limit when the end user should be warned that the size of the query is getting larger than is practical.
